requirements.txt
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
scikit-learn>=1.2.0
xgboost>=1.7.0
shap>=0.41.0
pyyaml>=6.0
jupyterlab>=3.6.0
pytest>=7.3.0